# docs/swagger.yaml
basePath: /api/v1/wallet
definitions:
...
  resp.CurrencyQuoteInfoResp:
    properties:
      base_currency:
        description: 基础货币
        type: string
      code:
        description: 标示
        type: string
      id:
        description: id
        type: integer
      price:
        description: 金额
        type: number
      quote_currency:
        description: 基础货币
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
...
  resp.ReadCoinDepositInfoResp:
    properties:
      coin_id:
        description: 代币id
        type: integer
      currency_id:
        description: 币种id
        type: integer
      min_withdrawal:
        description: 小于该金额不能提
        type: number
      money:
        description: 可提现金额
        type: number
      symbol:
        description: 币种代号
        type: string
      type:
        description: 标识 coin,token
        type: string
      withdrawal_fee:
        description: 提现手续费
        type: number
      withdrawal_fee_type:
        description: 手续费类型 fixed 按百分百比,percent 固定收取
        type: string
    type: object
  resp.ReadCoinInfoResp:
    properties:
      abi:
        type: string
      block_chain_id:
        description: 区块链名称
        type: string
      coin_id:
        description: 代币id
        type: integer
      confirm_count:
        description: 充值入帐的区块链确认数
        type: integer
      contract_address:
        description: 合约地址:如该是type=token，这里必须输入
        type: string
      currency_id:
        description: 货币id
        type: integer
      min_deposit:
        description: 最小充值金额，小于该金额不入账
        type: number
      min_withdrawal:
        description: 小于该金额不能提
        type: number
      name:
        description: 币种名称
        type: string
      status:
        description: 状态
        type: integer
      symbol:
        description: 币种代号
        type: string
      type:
        description: 标识 coin,token
        type: string
      withdrawal_fee:
        description: 提现手续费
        type: number
      withdrawal_fee_type:
        description: 手续费类型 fixed 按百分百比,percent 固定收取
        type: string
    type: object
  resp.ReadCoinListResp:
    properties:
      block_chain_id:
        description: 区块链名称
        type: string
      chain_code:
        description: 链代号
        type: string
      chain_name:
        description: 链名称
        type: string
      coin_id:
        description: 币种id
        type: integer
      confirm_count:
        description: 充值入帐的区块链确认数
        type: integer
      contract_address:
        description: 合约地址:如该是type=token，这里必须输入
        type: string
      currency_decimals:
        description: 代币小数点
        type: integer
      currency_id:
        description: 代币id
        type: integer
      currency_name:
        description: 代币名称
        type: string
      min_deposit:
        description: 最小充值金额，小于该金额不入账
        type: number
      min_withdrawal:
        description: 小于该金额不能提
        type: number
      name:
        description: 币种名称
        type: string
      status:
        description: 状态
        type: integer
      symbol:
        description: 币种代号
        type: string
      type:
        description: 标识 coin,token
        type: string
      updated_at:
        description: 更新时间
        type: string
      withdrawal_fee:
        description: 提现手续费
        type: number
      withdrawal_fee_type:
        description: 手续费类型 fixed 按百分百比,percent 固定收取
        type: string
    type: object
  resp.ReadCurrencyListResp:
    properties:
      currency_id:
        description: 币种id
        type: integer
      decimals:
        description: 小数点
        type: integer
      money:
        description: 可用金额
        type: number
      name:
        description: 名字
        type: string
      status:
        description: 状态：0开启;1:停用;2:删除
        type: integer
      symbol:
        description: 唯一标识
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
  resp.ReadCurrencyTransferListResp:
    properties:
      list:
        description: 可兑换列表
        items:
          $ref: '#/definitions/resp.ReadCurrencyListResp'
        type: array
      transfer:
        description: 兑换目标列表
        items:
          $ref: '#/definitions/resp.ReadCurrencyListResp'
        type: array
    type: object
...
info:
  contact: {}
  description: 测试
  license: {}
  title: 测试
  version: 0.0.1
paths:
...
  /coin/add:
    post:
      description: 增加代币
      parameters:
      - description: 币种id
        in: formData
        name: currency_id
        required: true
        type: integer
      - description: 代币代号
        in: formData
        name: symbol
        required: true
        type: string
      - description: 代币名称
        in: formData
        name: name
        required: true
        type: string
      - description: 链id
        in: query
        name: block_chain_id
        required: true
        type: integer
      - description: 币类型（coin,token）
        in: formData
        name: type
        required: true
        type: string
      - description: 确认数
        in: formData
        name: confirm_count
        required: true
        type: integer
      - description: 最小充值金额
        in: formData
        name: min_deposit
        required: true
        type: number
      - description: 最小提现金额
        in: formData
        name: min_withdrawal
        required: true
        type: number
      - description: 提现手续费
        in: formData
        name: withdrawal_fee
        required: true
        type: number
      - description: 手续费类型 1：百分比;2：固定类型
        in: formData
        name: withdrawal_fee_type
        required: true
        type: integer
      - description: 合约地址 如该是type=token，这里必须输入
        in: formData
        name: contract_address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 增加代币接口
      tags:
      - Coin 代币功能
  /coin/deposit:
    get:
      description: 读取代币提现信息
      parameters:
      - description: 代币id
        in: query
        name: coin_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.ReadCoinDepositInfoResp'
      security:
      - ApiKeyAuth: []
      summary: 读取代币提现信息接口
      tags:
      - Coin 代币功能
  /coin/info:
    get:
      description: 读取代币
      parameters:
      - description: 代币id
        in: query
        name: coin_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.ReadCoinInfoResp'
      security:
      - ApiKeyAuth: []
      summary: 读取代币接口
      tags:
      - Coin 代币功能
  /coin/list:
    get:
      description: 读取代币列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.ReadCoinListResp'
      security:
      - ApiKeyAuth: []
      summary: 读取代币列表接口
      tags:
      - Coin 代币功能
  /coin/remove:
    post:
      description: 删除代币
      parameters:
      - description: 代币id
        in: formData
        name: coin_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 删除代币接口
      tags:
      - Coin 代币功能
  /coin/status:
    post:
      description: 更新代币状态
      parameters:
      - description: 代币id
        in: formData
        name: coin_id
        required: true
        type: integer
      - description: 状态:0开启;1关闭
        in: formData
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 更新代币状态接口
      tags:
      - Coin 代币功能
  /coin/update:
    post:
      description: 更新代币
      parameters:
      - description: 币种id
        in: formData
        name: currency_id
        required: true
        type: integer
      - description: 代币id
        in: formData
        name: coin_id
        required: true
        type: integer
      - description: 代币代号
        in: formData
        name: symbol
        required: true
        type: string
      - description: 代币名称
        in: formData
        name: name
        required: true
        type: string
      - description: 链id
        in: formData
        name: block_chain_id
        required: true
        type: integer
      - description: 币类型（coin,token）
        in: formData
        name: type
        required: true
        type: string
      - description: 确认数
        in: formData
        name: confirm_count
        required: true
        type: integer
      - description: 最小充值金额
        in: formData
        name: min_deposit
        required: true
        type: number
      - description: 最小提现金额
        in: formData
        name: min_withdrawal
        required: true
        type: number
      - description: 提现手续费
        in: formData
        name: withdrawal_fee
        required: true
        type: number
      - description: 手续费类型 1：百分比;2：固定类型
        in: formData
        name: withdrawal_fee_type
        required: true
        type: integer
      - description: 合约地址 如该是type=token，这里必须输入
        in: formData
        name: contract_address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 更新代币接口
      tags:
      - Coin 代币功能
  /currency/add:
    post:
      description: 增加币种
      parameters:
      - description: 币种标识
        in: query
        name: symbol
        required: true
        type: string
      - description: 币种名称
        in: query
        name: name
        required: true
        type: string
      - description: 小数点数
        in: query
        name: decimals
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 增加币种接口
      tags:
      - currency 币种功能
  /currency/list:
    get:
      description: 读取币种列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.ReadCurrencyListResp'
      security:
      - ApiKeyAuth: []
      summary: 读取币种列表接口
      tags:
      - currency 币种功能
  /currency/quote:
    get:
      description: 货币报价
      parameters:
      - description: 报价货币
        in: query
        name: quote_currency
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.CurrencyQuoteInfoResp'
      security:
      - ApiKeyAuth: []
      summary: 货币报价接口
      tags:
      - currency 币种功能
  /currency/remove:
    post:
      description: 删除币种
      parameters:
      - description: 币种id
        in: query
        name: currency_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 删除币种接口
      tags:
      - currency 币种功能
  /currency/status:
    post:
      description: 更新币种状态
      parameters:
      - description: 币种id
        in: formData
        name: currency_id
        required: true
        type: string
      - description: 状态:0开启;1关闭
        in: formData
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 更新币种状态接口
      tags:
      - currency 币种功能
  /currency/transfer_list:
    get:
      description: 读取兑换币种列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.ReadCurrencyTransferListResp'
      security:
      - ApiKeyAuth: []
      summary: 读取兑换币种列表接口
      tags:
      - currency 币种功能
  /currency/update:
    post:
      description: 更新币种
      parameters:
      - description: 币种id
        in: query
        name: currency_id
        required: true
        type: string
      - description: 币种标识
        in: query
        name: symbol
        required: true
        type: string
      - description: 币种名称
        in: query
        name: name
        required: true
        type: string
      - description: 小数点数
        in: query
        name: decimals
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 更新币种接口
      tags:
      - currency 币种功能
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"